clarify mk_const_case_multiplier
Instead of rechecking every bit for constness at every level of
recursion, just identify the non-constant bits once up front and recurse
over the resulting list.

Besides making the recursive function easier to read, this allows
sorting the list of variable bits by index, to increase the chances of
simplifying the resulting if-then-else expressions.

Also, instead of modifying the list of bit expressions to introduce
constant values, just modify the final constant. This avoids O(2^n)
repeated calls to is_numeral which obviously should succeed but have to
be checked anyway. With this change, the types prove that the
intermediate values are numerals. Also I think the result is
asymptotically faster, but I don't want to think about it too hard.